chore: add jest-configurator for author-profile fix: utils/graphql.js now imports transpiled file
