+++ packages/author-profile/__tests__/android/author-profile.test.js
@@ -1,0 +1,3 @@
+import tests from "../author-profile-content.native";
+
+describe("AuthorProfile test on android", tests);
