--- packages/author-profile/jest.config.json
@@ -1,20 +1,0 @@
-{
-  "preset": "react-native",
-  "rootDir": "../../",
-  "coverageDirectory": "<rootDir>/packages/author-profile/coverage/native/",
-  "transformIgnorePatterns": [
-    "node_modules/(?!(react-native|react-apollo-temp|@times-components)/)"
-  ],
-  "testMatch": [
-    "<rootDir>/packages/author-profile/__tests__/**.native.test.js"
-  ],
-  "collectCoverageFrom": [
-    "**/packages/author-profile/*.js",
-    "!**/packages/author-profile/**.web.js",
-    "!**/packages/author-profile/fixture-generator.js",
-    "!**/packages/author-profile/*.fructose.*.js",
-    "!**/packages/author-profile/*.stories.js",
-    "!**/packages/author-profile/__tests__/**.test.js"
-  ],
-  "snapshotSerializers": ["enzyme-to-json/serializer"]
-}
