--- packages/author-profile/jest.config.web.json
@@ -1,27 +1,0 @@
-{
-  "rootDir": "../../",
-  "coverageDirectory": "<rootDir>/packages/author-profile/coverage/web/",
-  "transformIgnorePatterns": [
-    "node_modules/(?!(react-native|react-apollo-temp|@times-components)/)"
-  ],
-  "testEnvironment": "jsdom",
-  "testMatch": ["<rootDir>/packages/author-profile/__tests__/**.web.test.js"],
-  "moduleNameMapper": {
-    "react-native": "react-native-web"
-  },
-  "collectCoverageFrom": [
-    "**/packages/author-profile/*.js",
-    "!**/packages/author-profile/author-profile-item.js",
-    "!**/packages/author-profile/author-profile-list-provider.js",
-    "!**/packages/author-profile/fixture-generator.js",
-    "!**/packages/author-profile/*.fructose.*.js",
-    "!**/packages/author-profile/*.stories.js",
-    "!**/packages/author-profile/__tests__/**.test.js",
-    "!**/packages/author-profile/author-profile-error.js",
-    "!**/packages/author-profile/author-profile-listing-error.js",
-    "!**/packages/author-profile/author-profile-content.js"
-  ],
-  "moduleFileExtensions": ["web.js", "js", "json"],
-  "snapshotSerializers": ["enzyme-to-json/serializer"],
-  "setupFiles": ["raf/polyfill"]
-}
